requirements.txt
pika==0.12.0
pytz==2018.5
schedule==0.5.0
python-dotenv==0.7.1
requests==2.19.1
xmltodict==0.11.0